pyproject.toml
[project]
name = "ai-marketing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.14.0",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "mlflow>=2.22.0",
    "openai>=1.78.1",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
